Improve functional test
Include start page and login test case.
