+++ functional_tests/auth_dump.json
@@ -1,0 +1,1 @@
+[{"model": "auth.user", "pk": 1, "fields": {"password": "pbkdf2_sha256$180000$mbXpwHtZAnTO$QEHSTnO15CabCBuU4izxV7aJhtfx7sdEIDwUOjlrWAs=", "last_login": "2020-01-21T18:38:19.030Z", "is_superuser": true, "username": "admin", "first_name": "", "last_name": "", "email": "admin@mysite.com", "is_staff": true, "is_active": true, "date_joined": "2020-01-20T17:24:07.712Z", "groups": [], "user_permissions": []}}]
